% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{bootstrap}
\alias{bootstrap}
\title{Bagging}
\usage{
bootstrap(data)
}
\arguments{
\item{data}{the data list generated by \code{\link{buildTimeSeries}}}
}
\value{
The data list after bagging. The first two elements are the in-bag data (input and respond). The last two elements are the out-of-bag data (input and respond).
}
\description{
Generate bootstrap samples for determining the importance of prime implicant.
}
