% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{bubblesort}
\alias{bubblesort}
\title{Reorder the tree based on bubble sorting}
\usage{
bubblesort(tree)
}
\arguments{
\item{tree}{And/Or tree}
}
\value{
A sorted And/Or tree
}
\description{
A helper function in terms of the length of the tree
}
