% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{buildTimeSeries}
\alias{buildTimeSeries}
\title{Generate a list of data matrices}
\usage{
buildTimeSeries(network, numSeries, numPoints, noiseLevel, wildtype)
}
\arguments{
\item{network}{A network structure of class BooleanNetwork}

\item{numSeries}{The number of time series}

\item{numPoints}{The number of time points that is covered by each time series}

\item{noiseLevel}{The level of the nosie. e.g. 0.05 indiciates the value of a gene can be flipped with probability 0.05}

\item{wildtype}{a vector containing the wildtype expression values}
}
\value{
A list of data matrices containg the expression values of each node/gene. The row/column corresponds to time point/gene respectively. The second element refers to the input data and the third element refers to the respond data of the target gene.
}
\description{
Build the binary time series dataset matrix that used for network inference
}
\examples{
ngenes<-10
## k is the maximum number of genes
k<-5
## call generateRandomNKNetwork to generate a Boolean network
net1<-generateRandomNKNetwork(ngenes, k, topology="scale_free",simplify=TRUE,readableFunctions=TRUE)
datalist<-buildTimeSeries(network=net1,numSeries=10,numPoints=10,noiseLevel=0)

}
