% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{calImportance}
\alias{calImportance}
\title{Computing the importance of each prime implicant}
\usage{
calImportance(input, resp, forest)
}
\arguments{
\item{input}{A dataset matrix containing all gene expression. Each value corresponds to a gene expression of an input gene at a certain time point t.}

\item{resp}{A vector containing the gene expression of the target gene.}

\item{forest}{A set of And/Or trees, each tree is a list}
}
\value{
a vector includes the importances, the order of the importances corresponds to the order of prime implicants in the forest
}
\description{
A helper function for computing the importance of each prime implicant in a forest
}
\examples{
##Generate a random Boolean network
net1<-generateRandomNKNetwork(n=10, k=5, topology="scale_free",linkage="uniform",simplify=TRUE,readableFunctions=TRUE)

##Build the time-series data
datalist<-buildTimeSeries(net1,2,5,0)

##Select the first node as target gene and find 5 putative predictive And/Or trees using very simple parameters
datalist[[3]]<-datalist[[3]][,1]
forest<-lapply(rep(1,5),function(x){saalg(datalist,8,NULL,-2,-1,200)})

##To see all prime implicants in the forest
PIs<-unique(unlist(forest,recursive = F))
Importances<-calImportance(datalist[[2]],datalist[[3]],forest)
print(PIs)
print(Importances)

}
\references{
Müssel, Christoph, Martin Hopfensitz, and Hans A. Kestler. "BoolNet—an R package for generation, reconstruction and analysis of Boolean networks." Bioinformatics 26.10 (2010): 1378-1380.
}
\seealso{
\code{\link{generateRandomNKNetwork}}
}
