% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{changeName}
\alias{changeName}
\title{Output the name of final solution}
\usage{
changeName(singlename, ngenes)
}
\arguments{
\item{singlename}{A string represents a prime implicant}

\item{ngenes}{The number of nodes in the Boolean network; the genes in singlename should not be greater than ngenes}
}
\value{
The name of final solution
}
\description{
Change the name of prime implicants to corresponding "genes"
}
