% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{cyclic.covering}
\alias{cyclic.covering}
\title{Tree minimization helper function}
\usage{
cyclic.covering(mat, vec.primes)
}
\arguments{
\item{mat}{a matrix used for selecting prime implicants}
}
\description{
A helper function for minimizing the And/Or tree
}
