% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{findBF}
\alias{findBF}
\title{RFRE and Boolean function reconstruction}
\usage{
findBF(B, PIs, target, parameters, datalist, datasamples, seed)
}
\arguments{
\item{B}{The number of And/Or trees that used for generating a set of prime}

\item{PIs}{A list of prime implicants}

\item{target}{The integer indicating which node is the target node}

\item{parameters}{A vector includes 6 parameters; startT, endT and maxIter refer to the upper, lower temperature (on a log10 scale) and the maximum number of iterations used in simulated annealing algorithm
maxK represents the maximum number of input nodes of the target node (the size of the tree), it is required when prior knowledge, e.g. the in-degree is 8, is available. If such information is not known, then it can be set as a very large value, e.g. '.Machine$integer.max'
rate represents how many non-important PIs are removed in each recursion.
nodes represents the number of node in the Boolean network}

\item{datalist}{A data list generated by \link{buildTimeSeries} or \link{bootstrap}}

\item{datasamples}{The number of nodes in the Boolean network}
}
\value{
A final And/Or tree (i.e., a Boolean function)
}
\description{
A recursive feature reconstruction and elimination (RFRE) procedure for selecting a set of important prime implicants as features. Then final solution (i.e. Boolean function) is generated.
}
\examples{
## After we get a set of prime implicants, we use this function to generate the final Boolean function
## choose the same target gene
target<-3
datalist[[2]]<-generateData(PIs,datalist)
datalist[[3]]<-matrix(datalist[[3]][,target])
datasamples<-bootstrap(datalist)
datasamples$respinbag<-matrix(datasamples$respinbag)
datasamples$respoutbag<-matrix(datasamples$respoutbag)

parameters<-c(startT=2,endT=-1,maxIter=2000,maxK=8,rate=0.2,nodes=10)

## Identify the final Boolean function with 5 trees in the forest
findBF(5,PIs,target,parameters,datalist,datasamples,123)

}
\seealso{
\code{\link{generateRandomNKNetwork},\link{findPIs}}
}
