% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{findPIs}
\alias{findPIs}
\title{Find the important prime implicants}
\usage{
findPIs(B, datalist, datasamples, parameters, seed)
}
\arguments{
\item{B}{The number of And/Or trees that used for generating a set of prime}

\item{datalist}{A data list generated by \link{buildTimeSeries} or \link{bootstrap}}

\item{datasamples}{The number of nodes in the Boolean network}

\item{parameters}{A vector includes 6 parameters; startT, endT and maxIter refer to the upper, lower temperature (on a log10 scale) and the maximum number of iterations used in simulated annealing algorithm
maxK represents the maximum number of input nodes of the target node (the maximum number of leaves in an And/Or tree), it is required when prior knowledge, e.g. the in-degree is 8, is available. If such information is not known, then it can be set as a very large value, e.g. '.Machine$integer.max'
rate represents how many non-important PIs are removed in each recursion.
nodes represents the number of node in the Boolean network}
}
\value{
A set of important prime implicants. Or if only a few prime implicants are found, then a final Boolean function will be generated afterwards.
}
\description{
Given a time-series data, try to find the important prime implicants that are predictive for the gene expression level of a target gene
}
\examples{
## ngenes is the number of nodes in a Boolean network
library(BoolNet)
ngenes<-10
## k is the maximum number of genes
k<-5
## call generateRandomNKNetwork to generate a Boolean network
net1<-BoolNet::generateRandomNKNetwork(ngenes, k, topology="scale_free",simplify=TRUE,readableFunctions=TRUE)

## build the time-series data
datalist<-buildTimeSeries(network=net1,numSeries=10,numPoints=10,noiseLevel=0)

## select a target gene
target<-3
## Generate the bootstrap samples and oob samples according to the time-series data
datasamples<-bootstrap(datalist)
## respinbag and respoutbag save the expression values of the target node
datasamples$respinbag<-matrix(datasamples$respinbag[,target])
datasamples$respoutbag<-matrix(datasamples$respoutbag[,target])

## Initilize the parameters
parameters<-c(startT=2,endT=-1,maxIter=5000,maxK=8,rate=0.2,nodes=ngenes)

## Try to find the prime implicants
PIs<-findPIs(B=10,datalist,datasamples,parameters,123)


}
\seealso{
\code{\link{generateRandomNKNetwork}}
}
