% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{generateData}
\alias{generateData}
\title{Generate data matrices}
\usage{
generateData(PIs, datalist)
}
\arguments{
\item{PIs}{A list of prime implicants}

\item{datalist}{A list of data matrices that used to update a new data list}
}
\value{
A list of data matrices containg the expression values of PIs
}
\description{
A helper function for generating dataset in RFRE framework
}
