% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{growand}
\alias{growand}
\title{Subtree (And) growth based on individual nodes/genes}
\usage{
growand(tree, subtree, nodes, penalty = FALSE, dec)
}
\arguments{
\item{tree}{An And/Or tree}

\item{subtree}{A subtre, i.e., a prime implicant}

\item{nodes}{The number of nodes in the Boolean network}

\item{penalty}{TRUE or FALSE used to control the size of the tree}

\item{dec}{An argument refers to a specific movement}

\item{fast}{An argument used to indicate whether a sort is needed}
}
\value{
an And/Or tree
}
\description{
A helper function for running tree-growth while each leaf is a node/gene
}
