% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{growand2}
\alias{growand2}
\title{Subtree (And) growth based on PIs}
\usage{
growand2(tree, subtree, pis, maxK, currentnodes, allnodes,
  penalty = FALSE, dec)
}
\value{
an And/Or tree
}
\description{
A helper function for running tree-growth while each leaf is a prime implicant
}
