% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{growor}
\alias{growor}
\title{Subtree (Or) growth based on individual nodes/genes}
\usage{
growor(tree, nodes, penalty = FALSE, fast = TRUE)
}
\arguments{
\item{tree}{An And/Or tree}

\item{nodes}{The number of nodes in the Boolean network}

\item{penalty}{TRUE or FALSE used to control the size of the tree}

\item{fast}{An argument used to indicate whether a sort is needed}
}
\value{
an And/Or tree
}
\description{
A helper function for running tree-growth while each leaf is a node/gene
}
