% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{growtree}
\alias{growtree}
\title{Subtree growth based on individual nodes/genes}
\usage{
growtree(tree, maxK, nodes, penalty = FALSE, fast = TRUE)
}
\value{
an And/Or tree
}
\description{
A helper function for running tree-growth while each leaf is a node/gene
}
