% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{growtree2}
\alias{growtree2}
\title{Tree growth based on PIs}
\usage{
growtree2(tree, pis, maxK, currentnodes, allnodes, penalty = FALSE,
  fast = TRUE)
}
\value{
An And/Or tree
}
\description{
A helper function for running tree-growth while each leaf is a prime implicant
}
