% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{ia.samp}
\alias{ia.samp}
\title{Tree minimization helper function}
\usage{
ia.samp(n.pair, conj = 0)
}
\arguments{
\item{mat}{a matrix used for selecting prime implicants}
}
\description{
A helper function for minimizing the And/Or tree
}
