% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{minimization}
\alias{minimization}
\title{Tree minimization}
\usage{
minimization(tree, nodes)
}
\arguments{
\item{tree}{An And/Or tree}

\item{nodes}{the number of nodes in the Boolean network}
}
\value{
A minimized tree
}
\description{
A helper function for minimizing the And/Or tree
}
