% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{minimizePI}
\alias{minimizePI}
\title{Tree minimization helper function}
\usage{
minimizePI(mat)
}
\arguments{
\item{mat}{a matrix used for selecting prime implicants}
}
\description{
A helper function for minimizing the And/Or tree
}
