% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{mse}
\alias{mse}
\title{Find the better tree}
\usage{
mse(input, resp, tree1, tree2, nodes)
}
\arguments{
\item{input}{The input matrix in a data matrix}

\item{resp}{The respond in a data matrix}

\item{tree1}{An And/Or tree}

\item{tree2}{Another And/Or tree}

\item{nodes}{the number of nodes in the Boolean network}
}
\value{
the output difference between the output of tree1 and tree2. If a negative value is returned, then tree1 is better. If a positive value is returned, the tree2 is better. If zero is return, then tree1 and tree2 are equally good.
}
\description{
Given a data matrix with respond, try to find which tree is more predictive. The output of each tree is compared with the standard respond.
}
