% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{outputTree}
\alias{outputTree}
\title{Compute the output of the tree}
\usage{
outputTree(tree, input, nodes)
}
\arguments{
\item{tree}{An And/Or tree}

\item{input}{A data matrix, each row is a time point, each column is a node/gene.}

\item{nodes}{The number of nodes in the Boolean network}
}
\value{
a vector of the output of the tree
}
\description{
Compute the output of the tree according to the Boolean functions (And/Or tree) and the data matrix
}
