% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{reorderTree}
\alias{reorderTree}
\title{Reorder the tree}
\usage{
reorderTree(tree)
}
\arguments{
\item{tree}{And/Or tree}
}
\value{
A sorted And/Or tree
}
\description{
A helper function for sorting the And/Or tree
}
