% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{replaceName}
\alias{replaceName}
\title{Replace the names of new prime implicants with old prime implicants}
\usage{
replaceName(newnames, oldnames, allnodes)
}
\arguments{
\item{newnames}{A string representing new names of pis}

\item{oldnames}{A string representing old names of pis}

\item{allnodes}{The number of nodes in the Boolean network}
}
\value{
the correct names of pis
}
\description{
A helper function for replacing the names of new prime implicants (pis) and old prime implicants
}
\examples{
##old names
pis<-list(3,c(2,4),13)
oldnames<-sapply(pis,function(x){paste0(sort(x),collapse = "&")},simplify="array")
print(oldnames)

##new names "1&3" is expected as a combined string made up by "3&13", and assuming 10 nodes in total, then
newnames<-c("1&3")
print(newnames)#'
replaceName(newnames,oldnames,10)

newnames<-c("2&4")
##expect to be "2&4&13"
replaceName(newnames,oldnames,10)
}
