% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{rm.dom}
\alias{rm.dom}
\title{Tree minimization helper function}
\usage{
rm.dom(mat, col = FALSE, dom = TRUE)
}
\arguments{
\item{mat}{a matrix used for selecting prime implicants}
}
\description{
A helper function for minimizing the And/Or tree
}
