% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{saalg}
\alias{saalg}
\title{Tree growth based on individual nodes/genes}
\usage{
saalg(data, maxK, tree, initT, endT, iter)
}
\arguments{
\item{data}{The data list generated by \link{buildTimeSeries} or \link{bootstrap}}

\item{tree}{A initial tree}

\item{initT}{the uppper temperature}

\item{endT}{the lower temperature}

\item{iter}{the maximum number of iterations}

\item{maxk}{The maximum number of leaves/genes in the And/Or tree}
}
\value{
An And/Or tree
}
\description{
A helper function for running tree-growth based on simulated annealling algorithm
}
