% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{saalg2}
\alias{saalg2}
\title{Tree growth}
\usage{
saalg2(data, maxK, tree, initT, endT, iter, pis, allnodes)
}
\value{
An And/Or tree
}
\description{
A helper function for running tree-growth based on simulated annealling algorithm
}
