% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{selection}
\alias{selection}
\title{Tree growth parameters}
\usage{
selection()
}
\value{
a vector encoding the parameters used for tree growth
}
\description{
A helper function for tuning tree growth
}
