% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{changeName}
\alias{changeName}
\title{Output the name of final solution}
\usage{
changeName(singlename, ngenes)
}
\arguments{
\item{singlename}{a string represents a prime implicant}

\item{ngenes}{the number of nodes in the Boolean network}
}
\value{
The name of final solution
}
\description{
Change the name of prime implicants to corresponding "genes"
}
