% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{growand}
\alias{growand}
\title{Subtree (And) growth based on individual nodes/genes}
\usage{
growand(tree, subtree, nodes, penalty = FALSE, dec)
}
\value{
an And/Or tree
}
\description{
A helper function for running tree-growth while each leaf is a node/gene
}
