% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{growor2}
\alias{growor2}
\title{Subtree (Or) growth based on PIs}
\usage{
growor2(tree, pis, maxK, currentnodes, allnodes, penalty = FALSE,
  fast = TRUE)
}
\value{
T subtree
}
\description{
A helper function for running tree-growth while each leaf is a prime implicant
}
