% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEN.R
\name{saalg2}
\alias{saalg2}
\title{Tree growth}
\usage{
saalg2(data, maxK, tree, initT, endT, iter, pis, allnodes)
}
\arguments{
\item{data}{A data list generated by \link{buildTimeSeries} or \link{bootstrap}}

\item{maxK}{The size of the tree, or the maximum number of input genes; By defalt, maxK=8}

\item{initT}{the upper temperature (on a log10 scale)}

\item{endT}{the lower temperature (on a log10 scale)}

\item{iter}{the maximum number of iterations used in simulated annealing algorithm}

\item{pis}{A list of prime implicants}

\item{allnodes}{The number of genes in the Boolean network}
}
\value{
An And/Or tree
}
\description{
A helper function for running tree-growth based on simulated annealling algorithm
}
